requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
timm>=0.9.0
einops>=0.6.0
Pillow>=9.0.0
tqdm>=4.65.0
huggingface-hub>=0.16.0
matplotlib>=3.5.0
scikit-learn>=1.6.0
